PIM-3803 Remove js overrides
