+++ src/Pim/Bundle/DataGridBundle/Resources/config/requirejs.yml
@@ -17,10 +17,10 @@ config:
         pim/datagrid/column-form-listener:      pimdatagrid/js/datagrid/listener/column-form-listener
 
         oro/datagrid/action-launcher:           pimdatagrid/js/datagrid/action-launcher
-        oro/datagrid/abstract-action:           pimdatagrid/js/datagrid/action/oro-abstract-action
-        oro/datagrid/delete-action:             pimdatagrid/js/datagrid/action/oro-delete-action
+        oro/datagrid/abstract-action:           pimdatagrid/js/datagrid/action/abstract-action
+        oro/datagrid/delete-action:             pimdatagrid/js/datagrid/action/delete-action
         oro/datagrid/ajax-action:               pimdatagrid/js/datagrid/action/ajax-action
-        oro/datagrid/mass-action:               pimdatagrid/js/datagrid/action/oro-mass-action
+        oro/datagrid/mass-action:               pimdatagrid/js/datagrid/action/mass-action
         oro/datagrid/model-action:              pimdatagrid/js/datagrid/action/model-action
         oro/datagrid/navigate-action:           pimdatagrid/js/datagrid/action/navigate-action
         oro/datagrid/refresh-collection-action: pimdatagrid/js/datagrid/action/refresh-collection-action
@@ -28,7 +28,7 @@ config:
         oro/datagrid/actions-panel:             pimdatagrid/js/datagrid/actions-panel
         oro/datagrid/action-column:             pimdatagrid/js/datagrid/column/action-column
         oro/datagrid/body:                      pimdatagrid/js/datagrid/body
-        oro/datagrid/action-cell:               pimdatagrid/js/datagrid/cell/oro-action-cell
+        oro/datagrid/action-cell:               pimdatagrid/js/datagrid/cell/action-cell
         oro/datagrid/boolean-cell:              pimdatagrid/js/datagrid/cell/boolean-cell
         oro/datagrid/html-cell:                 pimdatagrid/js/datagrid/cell/html-cell
         oro/datagrid/date-cell:                 pimdatagrid/js/datagrid/cell/date-cell
