+++ src/Pim/Bundle/DataGridBundle/Resources/public/js/datagrid/action/mass-action.js
@@ -1,12 +1,43 @@
 /* global define */
-define(['underscore', 'oro/datagrid/mass-action'],
-function(_, MassAction) {
+define(['underscore', 'oro/messenger', 'oro/translator', 'oro/modal', 'oro/datagrid/abstract-action'],
+function(_, messenger, __, Modal, AbstractAction) {
     'use strict';
 
     /**
-     * Override abstract action to add the datalocale parameter
+     * Basic mass action class.
+     *
+     * @export  oro/datagrid/mass-action
+     * @class   oro.datagrid.MassAction
+     * @extends oro.datagrid.AbstractAction
      */
-    return MassAction.extend({
+    return AbstractAction.extend({
+        /** @property {Object} */
+        defaultMessages: {
+            confirm_title: __('Mass Action Confirmation'),
+            confirm_content: __('Are you sure you want to do this?'),
+            confirm_ok: __('Yes, do it'),
+            success: __('Mass action performed.'),
+            error: __('Mass action is not performed.'),
+            empty_selection: __('Please, select items to perform mass action.')
+        },
+
+        initialize: function(options) {
+            AbstractAction.prototype.initialize.apply(this, arguments);
+            this.route_parameters = _.extend(this.route_parameters, {gridName: this.datagrid.name, actionName: this.name});
+        },
+
+        /**
+         * Ask a confirmation and execute mass action.
+         */
+        execute: function() {
+            var selectionState = this.datagrid.getSelectionState();
+            if (_.isEmpty(selectionState.selectedModels) && selectionState.inset) {
+                messenger.notificationFlashMessage('warning', this.messages.empty_selection);
+            } else {
+                AbstractAction.prototype.execute.call(this);
+            }
+        },
+
         /**
          * Get action parameters
          *
@@ -35,8 +66,9 @@ function(_, MassAction) {
 
             return params;
         },
+
         /**
-         * Get extra parameters (sorters and )
+         * Get extra parameters (sorters and custom parameters)
          * @param {array}  params
          * @param {object} state
          *
@@ -52,6 +84,7 @@ function(_, MassAction) {
 
             return params;
         },
+
         /**
          * Get active sorters
          * @param {object} state
@@ -69,6 +102,7 @@ function(_, MassAction) {
 
             return result;
         },
+
         /**
          * Get active columns
          * @param {object} state
@@ -85,6 +119,24 @@ function(_, MassAction) {
             }
 
             return result;
+        },
+
+        _onAjaxSuccess: function(data, textStatus, jqXHR) {
+            this.datagrid.resetSelectionState();
+            AbstractAction.prototype._onAjaxSuccess.apply(this, arguments);
+        },
+
+        /**
+         * Get view for confirm modal
+         *
+         * @return {oro.Modal}
+         */
+        getConfirmDialog: function(callback) {
+            return new Modal({
+                title: this.messages.confirm_title,
+                content: this.messages.confirm_content,
+                okText: this.messages.confirm_ok
+            }).on('ok', callback);
         }
     });
 });
