--- src/Pim/Bundle/UIBundle/Resources/config/placeholders.yml
@@ -1,9 +1,0 @@
-placeholders:
-    requirejs_config_extend:
-        items:
-            require_js_config:
-                order: 110
-
-items:
-    require_js_config:
-        template: PimUIBundle::pim.requirejs.config.js.twig
