--- src/Pim/Bundle/UIBundle/Resources/views/pim.requirejs.config.js.twig
@@ -1,23 +1,0 @@
-require({
-    map: {
-        '*': {
-            'oro/datagrid/abstract-action':   'pimdatagrid/js/datagrid/action/abstract-action',
-            'oro/datagrid/mass-action':       'pimdatagrid/js/datagrid/action/mass-action',
-            'oro/datagrid/delete-action':     'pimdatagrid/js/datagrid/action/delete-action',
-            'oro/navigation/dotmenu/view':    'pimnavigation/js/navigation/dotmenu/view',
-            'oro/datagrid/action-cell':       'pimdatagrid/js/datagrid/cell/action-cell'
-        },
-        'pimdatagrid/js/datagrid/action/abstract-action': {
-            'oro/datagrid/abstract-action': 'oro/datagrid/abstract-action'
-        },
-        'pimdatagrid/js/datagrid/action/mass-action': {
-            'oro/datagrid/mass-action': 'oro/datagrid/mass-action'
-        },
-        'pimdatagrid/js/datagrid/action/delete-action': {
-            'oro/datagrid/delete-action': 'oro/datagrid/delete-action'
-        },
-        'pimnavigation/js/navigation/dotmenu/view': {
-            'oro/navigation/dotmenu/view': 'oro/navigation/dotmenu/view'
-        }
-    }
-});
